LayoutTests/imported/w3c/web-platform-tests/wasm/core/memory64/bulk64.wast.js-expected.txt


PASS #1 Reinitialize the default imports
PASS #2 Test that WebAssembly compilation succeeds (bulk64.wast:2)
PASS #3 Test that WebAssembly compilation succeeds (bulk64.wast:7)
FAIL #4 Test that WebAssembly compilation succeeds (bulk64.wast:45) assert_equals: expected false but got true
PASS #5 Test that WebAssembly compilation succeeds (bulk64.wast:117)
FAIL #6 Test that WebAssembly compilation succeeds (bulk64.wast:161) assert_equals: expected false but got true
PASS #7 Reinitialize the default imports
PASS #8 Test that WebAssembly compilation succeeds (bulk64.wast:2)
PASS #9 Test that WebAssembly instantiation succeeds (bulk64.wast:2)
PASS #10 Test that WebAssembly compilation succeeds (bulk64.wast:7)
PASS #11 Test that WebAssembly instantiation succeeds (bulk64.wast:7)
PASS #12 Run a WebAssembly test without special assertions (bulk64.wast:21)
PASS #13 Test that a WebAssembly code returns a specific result (bulk64.wast:22)
PASS #14 Test that a WebAssembly code returns a specific result (bulk64.wast:23)
PASS #15 Test that a WebAssembly code returns a specific result (bulk64.wast:24)
PASS #16 Test that a WebAssembly code returns a specific result (bulk64.wast:25)
PASS #17 Test that a WebAssembly code returns a specific result (bulk64.wast:26)
PASS #18 Run a WebAssembly test without special assertions (bulk64.wast:29)
PASS #19 Test that a WebAssembly code returns a specific result (bulk64.wast:30)
PASS #20 Test that a WebAssembly code returns a specific result (bulk64.wast:31)
PASS #21 Run a WebAssembly test without special assertions (bulk64.wast:34)
PASS #22 Run a WebAssembly test without special assertions (bulk64.wast:37)
PASS #23 Test that a WebAssembly code traps (bulk64.wast:40)
FAIL #24 Test that WebAssembly compilation succeeds (bulk64.wast:45) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 116: Data init_expr must produce an i32 at {loc} expected true got false
FAIL #25 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:55:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #26 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:58:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #27 Test that a WebAssembly code returns a specific result (bulk64.wast:62) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:61:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #28 Test that a WebAssembly code returns a specific result (bulk64.wast:63) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:64:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #29 Test that a WebAssembly code returns a specific result (bulk64.wast:64) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:67:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #30 Test that a WebAssembly code returns a specific result (bulk64.wast:65) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:70:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #31 Test that a WebAssembly code returns a specific result (bulk64.wast:66) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:73:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #32 Test that a WebAssembly code returns a specific result (bulk64.wast:67) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:76:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #33 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:79:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #34 Test that a WebAssembly code returns a specific result (bulk64.wast:71) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:82:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #35 Test that a WebAssembly code returns a specific result (bulk64.wast:72) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:85:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #36 Test that a WebAssembly code returns a specific result (bulk64.wast:73) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:88:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #37 Test that a WebAssembly code returns a specific result (bulk64.wast:74) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:91:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #38 Test that a WebAssembly code returns a specific result (bulk64.wast:75) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:94:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #39 Test that a WebAssembly code returns a specific result (bulk64.wast:76) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:97:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #40 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:100:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #41 Test that a WebAssembly code returns a specific result (bulk64.wast:80) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:103:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #42 Test that a WebAssembly code returns a specific result (bulk64.wast:81) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:106:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #43 Test that a WebAssembly code returns a specific result (bulk64.wast:82) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:109:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #44 Test that a WebAssembly code returns a specific result (bulk64.wast:83) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:112:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #45 Test that a WebAssembly code returns a specific result (bulk64.wast:84) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:115:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #46 Test that a WebAssembly code returns a specific result (bulk64.wast:85) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:118:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #47 Test that a WebAssembly code returns a specific result (bulk64.wast:86) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:121:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #48 Test that a WebAssembly code traps (bulk64.wast:89) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:124:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #49 Test that a WebAssembly code returns a specific result (bulk64.wast:92) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:127:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #50 Test that a WebAssembly code returns a specific result (bulk64.wast:93) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:130:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #51 Test that a WebAssembly code returns a specific result (bulk64.wast:94) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:133:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #52 Test that a WebAssembly code returns a specific result (bulk64.wast:95) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:136:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #53 Test that a WebAssembly code returns a specific result (bulk64.wast:96) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:139:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #54 Test that a WebAssembly code returns a specific result (bulk64.wast:97) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:142:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #55 Test that a WebAssembly code returns a specific result (bulk64.wast:98) assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_return@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:306:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:145:14
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #56 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:148:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #57 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:151:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #58 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:154:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #59 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:157:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #60 Test that a WebAssembly code traps (bulk64.wast:109) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:160:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #61 Test that a WebAssembly code traps (bulk64.wast:112) assert_true: expected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') assert_trap@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:263:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:163:12
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
PASS #62 Test that WebAssembly compilation succeeds (bulk64.wast:117)
PASS #63 Test that WebAssembly instantiation succeeds (bulk64.wast:117)
PASS #64 Run a WebAssembly test without special assertions (bulk64.wast:131)
PASS #65 Test that a WebAssembly code returns a specific result (bulk64.wast:132)
PASS #66 Test that a WebAssembly code returns a specific result (bulk64.wast:133)
PASS #67 Test that a WebAssembly code returns a specific result (bulk64.wast:134)
PASS #68 Run a WebAssembly test without special assertions (bulk64.wast:137)
PASS #69 Test that a WebAssembly code traps (bulk64.wast:140)
PASS #70 Test that a WebAssembly code returns a specific result (bulk64.wast:142)
PASS #71 Test that a WebAssembly code returns a specific result (bulk64.wast:143)
PASS #72 Run a WebAssembly test without special assertions (bulk64.wast:146)
PASS #73 Run a WebAssembly test without special assertions (bulk64.wast:147)
PASS #74 Test that a WebAssembly code traps (bulk64.wast:150)
PASS #75 Test that a WebAssembly code traps (bulk64.wast:153)
PASS #76 Run a WebAssembly test without special assertions (bulk64.wast:158)
FAIL #77 Test that WebAssembly compilation succeeds (bulk64.wast:161) assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.Module doesn't parse at byte 184: Data init_expr must produce an i32 at {loc} expected true got false
FAIL #78 Test that WebAssembly instantiation succeeds assert_true: unexpected instantiation error, observed TypeError: undefined is not an object (evaluating 'values[0].source') instance@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:201:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:214:18
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #79 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:217:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #80 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:220:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #81 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:223:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
FAIL #82 run assert_true: unexpected runtime error, observed TypeError: undefined is not an object (evaluating 'values[0].exports[name]') run@http://web-platform.test:8800/wasm/core/js/harness/async_index.js:242:24
bulk64_wast_js@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:226:4
global code@http://web-platform.test:8800/wasm/core/js/memory64/bulk64.wast.js:228:3 expected true got false
